Move BrowserPanel to controller package
