+++ src/main/java/seedu/address/controller/MainWindow.java
@@ -1,6 +1,5 @@
 package seedu.address.controller;
 
-import javafx.event.Event;
 import javafx.fxml.FXML;
 import javafx.scene.Node;
 import javafx.scene.Scene;
@@ -8,9 +7,7 @@ import javafx.scene.control.MenuItem;
 import javafx.scene.input.KeyCombination;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
-import javafx.scene.web.WebView;
 import javafx.stage.Stage;
-import seedu.address.browser.BrowserPanel;
 import seedu.address.events.controller.ExitAppRequestEvent;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
@@ -121,11 +118,11 @@ public class MainWindow extends UiPart {
     }
 
     void fillInnerParts() {
+        browserPanel = BrowserPanel.load(browserPlaceholder);
         personListPanel = PersonListPanel.load(primaryStage, getPersonListPlaceholder(), modelManager);
         resultDisplay = ResultDisplay.load(primaryStage, getResultDisplayPlaceholder());
         statusBarFooter = StatusBarFooter.load(primaryStage, getStatusbarPlaceholder(), config.getLocalDataFilePath());
         commandBox = CommandBox.load(primaryStage, getCommandBoxPlaceholder(), parser, resultDisplay, modelManager);
-        browserPanel = BrowserPanel.load(browserPlaceholder);
     }
 
     private AnchorPane getCommandBoxPlaceholder() {
